Persist tab bar selection even when closing app, project organization
