+++ Course/CourseApp.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct CourseApp: App {
     var body: some Scene {
         WindowGroup {
-            AccountView()
+            ContentView()
         }
     }
 }
