+++ Course/Navigation/TabBar.swift
@@ -9,50 +9,32 @@ import SwiftUI
 
 struct TabBar: View {
 
-    @State var selectedTab: Tab = .learnNow
+    @AppStorage("selectedTab") var selectedTab: Tab = .learnNow
     @State var color: Color = .teal
     @State var tabItemWidth: CGFloat = 0
 
     var body: some View {
-        ZStack(alignment: .bottom) {
-
-            Group {
-                switch selectedTab {
-                case .learnNow:
-                    ContentView()
-                case .explore:
-                    AccountView()
-                case .notifications:
-                    AccountView()
-                case .library:
-                    AccountView()
-                }
-            }
-            .frame(maxWidth: .infinity, maxHeight: .infinity)
-
-            HStack {
-                buttons
-            }
-
-            .padding(.horizontal, 8)
-            .padding(.top, 14)
-            .frame(height: 88, alignment: .top)
-            .background(
-                .ultraThinMaterial,
-                in: RoundedRectangle(cornerRadius: 34, style: .continuous)
-            )
-            .background(
-                background
-            )
-            .overlay(
-                overlay
-            )
-
-            .strokeStyle(cornerRadius: 34)
-            .frame(maxHeight: .infinity, alignment: .bottom)
-            .ignoresSafeArea()
+        HStack {
+            buttons
         }
 
+        .padding(.horizontal, 8)
+        .padding(.top, 14)
+        .frame(height: 88, alignment: .top)
+        .background(
+            .ultraThinMaterial,
+            in: RoundedRectangle(cornerRadius: 34, style: .continuous)
+        )
+        .background(
+            background
+        )
+        .overlay(
+            overlay
+        )
+
+        .strokeStyle(cornerRadius: 34)
+        .frame(maxHeight: .infinity, alignment: .bottom)
+        .ignoresSafeArea()
     }
 
     var buttons: some View {
